* ser_posix.c, ser_win32.c, serial.h: Change baud from int to long to avoid a 16-bit int overflow.
git-svn-id: svn://svn.savannah.nongnu.org/avrdude/trunk@291 81a1dc3b-b13d-400b-aceb-764788c761c2

+++ avrdude/ser_win32.c
@@ -20,14 +20,16 @@
 /* $Id$ */
 
 /*
- * Posix serial interface for avrdude.
+ * Native Win32 serial interface for avrdude.
  */
+
+#include "serial.h"
 
 extern char *progname;
 
 #if 0
 
-int serial_open(char * port, int baud)
+int serial_open(char * port, long baud)
 {
   return fd;
 }
